Importind the complete project
The project is based on html, css, javascript and bootstrap technologies. The website is like social media platform. User can sign in and sign Out into the website, Create the post and fill the personal details in the Google form and also Chat with the other people. They can even subscribe our website. This is done in bootstrap.
If in case any queries they can connect us with our other social media platforms like facebook, Twitter, Instagram and LinkedIn